Move Exie controls to system settings
